Add check for retain-certs config before clearing certs

+++ src/charm.py
@@ -241,9 +241,10 @@ class ZooKeeperCharm(CharmBase):
         self.on[self.restart.name].acquire_lock.emit()
 
     def _on_certificate_relation_departed(self, event: EventBase):
-        for file in ["ca", "cert", "key", "common_name"]:
-            if os.path.exists(f"{TLS_STORE_DIR}/{file}"):
-                os.remove(f"{TLS_STORE_DIR}/{file}")
+        if not self.config["retain-certs"]:
+            for file in ["ca", "cert", "key", "common_name"]:
+                if os.path.exists(f"{TLS_STORE_DIR}/{file}"):
+                    os.remove(f"{TLS_STORE_DIR}/{file}")
 
 
 if __name__ == "__main__":
